fix(hooks): treat undefined as null in getTypeOf
Happens to remove the schema dependency of the function, since we no
longer are returning the preferrred type for undefined values.

This creates better behavior out of assertType, which will now return
the default for the schema when the value is undefined (which should
still be null if null types are allowed).